% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OKcheck-package.R
\docType{package}
\name{OKcheck-package}
\alias{OKcheck}
\alias{OKcheck-package}
\title{OKcheck: Tools to facilitate checking of data from National Surveillance Programmes}
\description{
Provide tools to facilitate checking of data from national surveillance programmes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NorwegianVeterinaryInstitute/OKcheck}
  \item Report bugs at \url{https://github.com/NorwegianVeterinaryInstitute/OKcheck/issues}
}

}
\author{
\strong{Maintainer}: Petter Hopp \email{Petter.Hopp@vetinst.no} (\href{https://orcid.org/0000-0002-8695-0378}{ORCID})

Other contributors:
\itemize{
  \item Norwegian Veterinary Institute [copyright holder]
}

}
\keyword{internal}
