% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_for_checking.R
\name{write_data_for_checking}
\alias{write_data_for_checking}
\title{Write table with data that should be checked}
\usage{
write_data_for_checking(data, heading)
}
\arguments{
\item{data}{data.frame with data that should be checked by the laboratory.}

\item{heading}{Table title and text.}
}
\value{
If no lines to check an empty document. Else, a table with potential
mistakes in PJS. \code{standardize_columns} is used to standardize column
names.
}
\description{
Write a table with data that should be checked by those responsible
for entering the data.
}
\details{
The data checks PJS-data for potential mistakes. If data with potential
mistakes are identified, these are reporter in a table. If no mistakes are
found, no report is created.\preformatted{The heading should explain why the data needs to be checked so that the
laboratory should know what to change in the source data, i.e. PJS.
}
}
