% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_PJScodes.R
\name{count_PJScodes}
\alias{count_PJScodes}
\title{Count used PJS-codes}
\usage{
count_PJScodes(
  PJSdata,
  variable,
  accepted = NULL,
  translation_table = PJS_codes_2_text
)
}
\arguments{
\item{PJSdata}{Data frame with data from PJS.}

\item{variable}{The variabel in the data frame that should be checked.}

\item{translation_table}{The translation table for PJS-codes, see details.}
}
\value{
A data frame with the used codes, their description and the number of
rows where the codes have been used.
}
\description{
Count used PJS-codes and marks whether they are accepted or not.
}
\details{
The purpose is to give an overview of the PJS-codes used. PJS-codes
that are included in the control routines as accepted codes are marked. The
description text is added to all used codes to facilitate checking if the
codes shoud be accepted or not.

The \code{translation_table} for PJS-codes neeeds to be imported. The function
\code{NVIdb::read_PJS_code_2_text} should be used, see example.
}
\examples{
\dontrun{
# Reads translation table for PJS-codes
PJS_codes_2_text <- NVIdb::read_PJS_codes_2_text()

ktr <- count_PJScodes(PJSdata = PJSdata,
                      variable = "metodekode",
                      accepted = c("010001", "010002"),
                      translation_table = PJS_codes_2_text)
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
