% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_table_if_data.R
\name{knit_table_if_data}
\alias{knit_table_if_data}
\title{Knit table if data to report}
\usage{
knit_table_if_data(data, heading)
}
\arguments{
\item{data}{data.frame with or without data.}

\item{heading}{Table title and caption.}
}
\value{
A knitted table formatted by kable for output from rmarkdown. If no
data in the data source, an empty document.
}
\description{
Knit a table using \code{kable} if there are data in the data frame.
}
\details{
Knits a table within an rmarkdown document if there are data in the data
source.

The column names are standardized using \code{standardize_columns}. The
output also includes a table heading.

This is primary used in rmarkdown reports for checking PJS-data for
potential mistakes. If data with potential mistakes are identified, these
are reported in a table. If no mistakes are found, no report is created.

The heading should explain why the data needs to be checked so that the
laboratory should know what to change in the source data, i.e. PJS.
}
